trying to get app down to just... app
