+++ memgpt/server/constants.py
@@ -1,6 +1,3 @@
 # WebSockets
 WS_DEFAULT_PORT = 8282
 WS_CLIENT_TIMEOUT = 30
-
-# REST
-REST_DEFAULT_PORT = 8283
